organized files into directories
